리팩토링: 경고 제거
프로젝트의 경고를 제거한다.

+++ TestBed/Classes/ScanView/GLView.mm
@@ -63,7 +63,7 @@
  @param depth 깊이값
  */
 - (void)setDepth:(NSInteger)depth {
-    _esContext->_depth = depth;
+    _esContext->_depth = (int)depth;
 }
 
 /**
